server_pandas/requirements.txt
appdirs==1.4.4
asn1crypto==0.24.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cloud-init==19.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.1.4
DateTime==4.3
defer==1.0.6
distlib==0.3.1
distro-info===0.18ubuntu0.18.04.1
filelock==3.0.12
Flask==1.1.2
html5lib==1.1
httplib2==0.9.2
idna==2.6
importlib-metadata==4.0.1
importlib-resources==5.1.2
isc==2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
language-selector==0.1
MarkupSafe==1.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==2.0.6
pandas==1.1.5
ply==3.11
pybithumb==1.0.21
pygobject==3.26.1
PyJWT==1.5.3
pyserial==3.4
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
python-dotenv==0.17.1
pytz==2021.1
PyYAML==3.12
redis==3.5.3
requests==2.18.4
requests-unixsocket==0.1.5
six==1.11.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
virtualenv==20.4.6
webencodings==0.5.1
websockets==9.0.1
Werkzeug==1.0.1
xkit==0.0.0
zipp==3.4.1
zope.interface==5.4.0
